# perf: speed up worlds list query